requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Werkzeug==2.3.7
Flask-Limiter==2.7.0
python-dotenv>=1.0.0
pyodbc>=5.0.0
